requirements.txt
alembic==0.8.4
appnope==0.1.0
arrow==0.7.0
awesome-slugify==1.6.5
beautifulsoup4==4.4.1
bunch==1.0.1
cached-property==1.3.0
cairocffi==0.7.2
CairoSVG==1.0.19
cairotft==0.1.2
cffi==1.3.1
cssselect==0.9.1
dataset==0.6.1
decorator==4.0.6
discover==0.4.0
docopt==0.6.2
ERPpeek==1.6.3
et-xmlfile==1.0.1
gnureadline==6.3.3
html5lib==0.9999999
ipython==4.0.3
ipython-genutils==0.1.0
jdcal==1.2
Jinja2==2.8
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
mysqlclient==1.3.7
normality==0.2.4
nose==1.3.7
numpy==1.10.2
openpyxl==2.3.2
pandas==0.17.1
path.py==8.1.2
petl==1.1.0
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5
py==1.4.31
pycparser==2.14
Pyphen==0.9.2
pytest==2.8.4
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.7
PyYAML==3.11
regex==2015.11.22
requests==2.9.1
selenium==2.48.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.9
sqlparse==0.1.18
tinycss==0.3
traitlets==4.1.0
Unidecode==0.4.18
WeasyPrint==0.24
wheel==0.24.0
xlrd==0.9.4
